java core: Implement base class serialization.
This splits out field serialization from the top-level .serialize() into
a protected method, and makes that method responsible for calling its
equivalent on the supertype if there is one. This should give us the
correct paths through {read,write}{Struct,Base}{Begin,End}.

The new members, .serializeFields and .deserializeFields, are strong
candidates for the static inner helper we've been talking about.
Creating it now would make this change larger than I would like it to
be.